Assistant checkpoint: Fix PDF handling: restore reading and improve export
Assistant generated file changes:
- client/src/lib/textExtractor.ts: Restore PDF content extraction
- client/src/lib/pdf.ts: Improve PDF generation to match translation format

---

User prompt:

See, this is what i'm talking about, you broke something that was working just fine, that I don't think we even needed to touch:

%PDF-1.5
%����
26 0 obj
<</Linearized 1/L 95589/O 28/E 79341/N 5/T 95246/H [ 491 219]>>
endobj

42 0 obj
<</DecodeParms<</Columns 5/Predictor 12>>/Filter/FlateDecode/ID[<32E83D8067F784E28AB7707BECDD1F51><84A079DBFA3F48478B86713BAF49FE0B>]/Index[26 25]/Info 25 0 R/Length 87/Prev 95247/Root 27 0 R/Size 51/Type/XRef/W[1 3 1]>>stream
h�bbd```b``�"A$��f&���`�D2'�٦`�L
��/`�l� ������i1Ж�`qF��s�}m
endstream
endobj

The above is the 'source text' display now, when we're uploading a PDF, it not longer displayes properly.

Why are we messing with the upload portion of it? That wasn't the issue, the issue is with the creation of the pdf that we export. We need to create the PDF, using the exact same font, and format as what is shown in the 'translation' text box